requirements.txt
streamlit==1.28.1
streamlit-authenticator>=0.4.0
pandas==2.0.3
numpy==1.24.3
psycopg2-binary==2.9.7
sqlalchemy==2.0.19
alembic==1.11.1
python-dotenv==1.0.0
reportlab==4.0.4
weasyprint==59.0
Pillow==10.0.0
email-validator==2.0.0
python-multipart==0.0.6
jinja2==3.1.2
apscheduler==3.10.1
boto3==1.28.3
sendgrid==6.10.0
pytz==2023.3
bcrypt==4.0.1
cryptography==41.0.3
pydantic==2.1.1
python-dateutil==2.8.2
plotly==5.15.0
openpyxl==3.1.2
xlsxwriter==3.1.2
openai>=1.0.0

# MCP Server Dependencies
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-auth==2.23.4